% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introjs.R
\name{introjs}
\alias{hintjs}
\alias{introjs}
\alias{introjsUI}
\title{Initiate intro.js}
\usage{
introjs(session)

hintjs(session)

introjsUI(includeOnly = FALSE)
}
\arguments{
\item{session}{the Shiny session object (from the server function of the Shiny app)}

\item{includeOnly}{Only include introjs files. For users who will write their own javascript}
}
\description{
Iniates an introduction via the intro.js library
}
\note{
\code{introjsUI} must be present in UI for \code{introjs} to work.
 Also requires their to be \code{introBox} in UI.
}
\examples{
\dontrun{
library(rintrojs)
library(shiny)
ui <- shinyUI(fluidPage(
  introjsUI(),
  mainPanel(
    introBox(
      tableOutput("mtcars"),
      data.step = 1,
      data.intro = "This is the table"
    ),
    introBox(
      actionButton("btn","Intro"),
      data.step = 2,
      data.intro = "This is the button"
    )
  )))
server <- shinyServer(function(input, output, session) {
  output$mtcars <- renderTable({
    head(mtcars)
  })
  observeEvent(input$btn,
               introjs(session))
})
# Run the application
shinyApp(ui = ui, server = server)
}
}
\seealso{
\code{introBox}
}

